panes/LocationPane.jsx -> panes/AddLocationPane.jsx
and adding logic for pending locations.
